refactor(web): deduplicate dashboard type definitions into a shared module
